Move IO functionality into io.py
