+++ pyaldata/__init__.py
@@ -4,3 +4,5 @@ from .interval import *
 from .regression import *
 from .dim_reduction import *
 from .movement_onset_and_peak import *
+from .data_cleaning import *
+from .io import *
