+++ pyaldata/dim_reduction.py
@@ -26,7 +26,7 @@ def fit_dim_reduce_model(trial_data, model, signal, train_trials=None, fit_kwarg
      Example
      -------
          from sklearn.decomposition import PCA
-         pca_dims = -5
+         pca_dims = 5
          pca = fit_dim_reduce_model(trial_data, PCA(pca_dims), 'M1_rates')
      """
      if fit_kwargs is None:
